+++ packages/gamut/src/PopoverContainer/PopoverContainer.tsx
@@ -40,6 +40,7 @@ export const PopoverContainer: React.FC<PopoverContainerProps> = ({
   targetRef,
   allowPageInteraction,
   closeOnViewportExit = false,
+  focusOnProps,
   ...rest
 }) => {
   const popoverRef = useRef<HTMLDivElement>(null);
@@ -265,6 +266,7 @@ export const PopoverContainer: React.FC<PopoverContainerProps> = ({
   const content = (
     <FocusTrap
       allowPageInteraction={inline || allowPageInteraction}
+      focusOnProps={focusOnProps}
       onClickOutside={handleClickOutside}
       onEscapeKey={onRequestClose}
     >